backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.1
python-dotenv==1.0.1
pydantic-settings==2.7.1
httpx==0.28.1
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
google-auth==2.38.0
python-multipart==0.0.20
langchain>=0.3.14
langgraph>=0.2.60
langchain-google-genai>=2.0.8
pdfplumber>=0.11.4
python-docx>=1.1.2
thefuzz[speedup]>=0.22.1
google-api-python-client>=2.159.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
beautifulsoup4>=4.12.3
langchain-groq>=0.3.0
websockets>=14.0
